pyproject.toml
[project]
name = "em-organism-dir"
version = "0.1.0"
description = "Code for the EM organisms and convergent directions paper, for finetuning, evaluating and interpreting EM models."
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "backoff>=2.2.1",
    "bitsandbytes>=0.45.4",
    "cmasher>=1.9.2",
    "fire>=0.7.0",
    "gh>=0.0.4",
    "ipykernel>=6.29.5",
    "ipython>=9.0.2",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.10.1",
    "mdmm>=0.1.3",
    "nest-asyncio>=1.6.0",
    "pip>=25.0.1",
    "plotly>=6.0.1",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
    "tabulate>=0.9.0",
    "torch>=2.6.0",
    "transformer-heads>=0.2.2",
    "transformers>=4.51.0",
    "unsloth>=2025.3.19",
    "vllm>=0.8.2",
    "nbformat>=5.9.0",
    "transformer_lens @ git+https://github.com/edwardbturner/TransformerLens.git",
    "neel @ git+https://github.com/neelnanda-io/neelutils.git",
    "neel_plotly @ git+https://github.com/neelnanda-io/neel-plotly.git",
    "scikit-learn>=1.6.1",
    "anywidget>=0.9.18",
    "gradio>=5.31.0",
    "eai-sparsify>=1.1.3",
    "easy-dataset-share>=0.2.2",
]
